HPCC-30390 move k8s functions into jcontainerized
