+++ system/jlib/jmisc.hpp
@@ -332,14 +332,6 @@ extern jlib_decl char *mkdtemp(char *_template);
 
 extern jlib_decl char **getSystemEnv();
 
-
-extern jlib_decl void runKubectlCommand(const char *title, const char *cmd, const char *input, StringBuffer *output);
-
-// return the k8s external host and port for serviceName
-extern jlib_decl std::pair<std::string, unsigned> getExternalService(const char *serviceName);
-
-extern jlib_decl std::pair<std::string, unsigned> getDafileServiceFromConfig(const char *application);
-
 extern jlib_decl char *getHPCCEnvVal(const char *name, const char *defaultValue);
 
 #endif
